+++ src/JWT/Action/FetchGooglePublicKeys.php
@@ -9,23 +9,34 @@ use Kreait\Firebase\JWT\Value\Duration;
 
 final class FetchGooglePublicKeys
 {
+    /** @deprecated 1.15.0 */
     const DEFAULT_URL = 'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com';
+
+    const DEFAULT_URLS = [
+        'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com',
+        'https://www.googleapis.com/oauth2/v1/certs',
+    ];
+
     const DEFAULT_FALLBACK_CACHE_DURATION = 'PT1H';
 
-    /** @var string */
-    private $url = self::DEFAULT_URL;
+    /** @var array<int, string> */
+    private $urls;
 
     /** @var Duration */
     private $fallbackCacheDuration;
 
-    private function __construct()
+    /**
+     * @param array<array-key, string> $urls
+     */
+    private function __construct(array $urls, Duration $fallbackCacheDuration)
     {
-        $this->fallbackCacheDuration = Duration::fromDateIntervalSpec(self::DEFAULT_FALLBACK_CACHE_DURATION);
+        $this->urls = \array_values($urls);
+        $this->fallbackCacheDuration = $fallbackCacheDuration;
     }
 
     public static function fromGoogle(): self
     {
-        return new self();
+        return new self(self::DEFAULT_URLS, Duration::fromDateIntervalSpec(self::DEFAULT_FALLBACK_CACHE_DURATION));
     }
 
     /**
@@ -33,31 +44,39 @@ final class FetchGooglePublicKeys
      */
     public static function fromUrl(string $url): self
     {
-        $action = new self();
-        $action->url = $url;
-
-        return $action;
+        return new self([$url], Duration::fromDateIntervalSpec(self::DEFAULT_FALLBACK_CACHE_DURATION));
     }
 
     /**
      * A response from the Google APIs should have a cache control header that determines when the keys expire.
      * If it doesn't have one, fall back to this value.
      *
-     * @param Duration|DateInterval|string| int $duration
+     * @param Duration|DateInterval|string|int $duration
      */
     public function ifKeysDoNotExpireCacheFor($duration): self
     {
         $duration = Duration::make($duration);
 
-        $action = new self();
+        $action = clone $this;
         $action->fallbackCacheDuration = $duration;
 
         return $action;
     }
 
+    /**
+     * @deprecated 1.15.0
+     */
     public function url(): string
     {
-        return $this->url;
+        return $this->urls[0];
+    }
+
+    /**
+     * @return array<int, string>
+     */
+    public function urls(): array
+    {
+        return $this->urls;
     }
 
     public function getFallbackCacheDuration(): Duration